Fix dangerous sprintf() calls [CBLUR-57]
Introduces a replacement sprintf()-like function sdprintf()
that will use the provided buffer if it is of sufficient size,
or else allocate a new one that is.

Returned pointer is wrapped in a 'smart' pointer class that
manages the lifetime of the buffer.

Change-Id: Ifef4a6ec76b871a9027f49e9ecde04b4b03f9665

CoreNEURON Repo SHA: BlueBrain/CoreNeuron@62d1b8d922e8ef55f6cbaaedc251c5cb24af3508