Many existing gui tools do not work with Python Sections. That is, a Python Section name is useless as part of a hoc statement. To partially alleviate this problem, all sections, hoc or python created, have an internal name accepted by the hoc interpreter that begins with __nrnsec_0x and end with its hexadecimal memory location. In Python, nrn.Section.hoc_internal_name() returns this name. That name can serve as the section reference in hoc anywhere the normal section syntax is allowed. When hoc parses the __nrnsec_0x... token or python does its get attribute for h.__nrnsec_0x... then, if the relevant section does not exist, either the error message "Invalid internal section name: __nrnsec_0x..." or "Section associated with internal name does not exist: __nrnsec_0x..." will be generated.
The const char* secname(Section*) that is used everywhere
to return the name of a section has been modified to return the
hoc_internal_name for python sections that were not created with a
name='...' named argument (instead of, e.g, 'PySec_0x7f73f2650558')

To prevent segmentation violations if the user inadvertently uses a
__nrnsec_0x... token that results in an invalid memory address, all
sections are allocated and freed from a SectionPool of 1000 sections which
doubles in size, without freeing earlier pool memory, in case of overflow.
int nrn_is_valid_section_pr(void* vp) safely returns 0 if vp does not point
to a possible Section in the pool.
.