--cell_permute=2 and missing --gpu not allowed when compiled with _OPENACC
--voltage=1000 means set the first arg of nrn_finitialize to 0. Otherwise
it is set to 1 and the voltage is use to initialize all compartments (before
finitialize calls the mod file INITIAL blocks). Note that when the first
arg is 0, then the initial voltages are the same as specified in the dat file
but the INITIAL blocks are still called.

CoreNEURON Repo SHA: BlueBrain/CoreNeuron@de5a5576b8435b769fd73ce3dc0e671139fe243d