python_already_loaded() can return NULL even when python library is loaded.
On recent mac osx (eg 10.12 Sierra) when python is launched,
python_already_loaded returns NULL if the python library is located in
a system library (an effect of System Integrity Protection?) unless
the full path to the dylib is used. The result of the NULL handle was that
the nrnpython interface library was not loaded. Since we know it is loaded
in this circumstance, it was sufficient to go ahead and load
libnrnpython...dylib anyway.