Support python2 and python3 versions of libnrnpython in same installer.
configure option --with-nrnpython-only in conjunction with, e.g.,
--with-nrnpython=dynamic --with-pyexc=python3
will build only in the src/nrnpython folder and install libnrnpython35.so
and setup.py for python3. This must be subsequent to full install without
that configure option and --with-pyexc=python2 in order to install
everything outside the nrnpython folder that is needed for
building libnrnpython35.so and successfully executing setup.py for python3.

Note that if python is launched, the proper libnrnpython will be loaded
based on nrn_is_python_extension which has the value of the launched
sys.version without the '.'.
If nrniv is launched with the python library specified by the
NRN_PYLIB environment variable, the proper libnrnpython will be determined
from the name of the python library. If nrniv is launched and the default
python library is loaded, a dynamic load of libnrnpython27.so will be attempted
first.