Remove early return from nrnhash key generation
The hash generation had a return prior to executing numeric operations.  This seems unintentional
since the following code lines were never able to be executed.

This was noticed during bbcorewrite where the key comes from a void pointer.  The early return
meant that the hashes used were not evenly distributed.