modified basicRxD3D test to allow multiple runs (#74)
Now setting the default dimension to 3 instead of any specific Section's dimension. This means that when new Sections are created they will default to 3D simulation so there will be no conflict with a pre-existing 3D matrix.

The underlying issue is that NEURON currently requires nodes to be grouped with like nodes (e.g. all 1D then all 3D) and thus changing things after the matrices are constructed is fragile.