cvode.statename supported by nonvint_block_supervisor
set_nonvint_block takes a second callable (or NULL) with the signature
char*(*new_nrn_nonvint_statename)(int index, int tid)
The c free() will be called on the returned string by the caller so
const char* or strings that need to be persistent should be strdup
and returned by the callback.

The statename callback from the nonvint_block_supervisor is the 11th
callback and takes an index arg and must return a Python string.