Only binary special's, linking against nrnivmech
When NRNMECH_DLL_STYLE_TRUE was true, special was not a binary
anymore, but a shell script.
With this patch we are creating a binary special with the makefile
which links against libnrnmech.
Using libtool, we install the lib to the intended rpath, with the
additional benefit that libtool no longer creates special in .libs

This is an initial version that requires further testing