The cmake options NRNPYTHON_DYNAMIC and INSTALL_NEURON_MODULE work.
Default python3 falling back to python.

If NRNPYTHON_DYNAMIC is in effect the NRNPYTHON_SRC is no longer
in libnrniv and libnrnpython2.so or libnrnpython3.so are loaded as
needed depending on the python version that is being used.

If INSTALL_NEURON_MODULE is YES
Setup.py for each python executable in the NRNPYTHON_DYNAMIC list
(or the default python executable if NRNPYTHON_DYNAMIC is YES or NO) is
used to install the neuron modules using --home=${CMAKE_INSTALL_PREFIX} .
If the value of INSTALL_NEURON_MODULE is neither YES nor NO, the module
is installed with <python> setup.py install <value of INSTALL_NEURON_MODULE>