Now allow h.delete_section(sec=pysec) for PythonSection.
Note that the PythonSection retains its reference to the now invalid
NEURON internal Section and former, now zombie PythonSection
is freed when the PythonSection refcnt goes to 0. Of course the internal
invalid NEURON Section is freed when it's refcount goes to 0 (in usual
circomstances, when the zombie PythonSection is destroyed.)
SectionRef.unname() and SectionRef.rename() do nothing and return 0.0
when SectionRef is referenceing a PythonSection.