Pyiter3 (#457)
* Reimplement segments of section and mechanisms of segment iterators.

This supports iter(soma) and iter(soma(.5)).

* Section and Segment iterators mostly follow the PyList impl pattern.

* nrn.VarOfMechIter interates over all nrn.RangeVar of a nrn.Mechanism

RangeVar extended from arrays to handle scalars with syntax
nrn.RangeVar[0] for assignment and evaluation.

* Test for section, segment, mechanism, rangvar iterators

* Fix memory leak due to extra reference of Section.
E.g. a python section that becomes currently accessed section.

* Re-implement common iterator for SectionList and h.allsec().
Avoid using freed memory when iterator variable starts out as the
only reference to the second item of section_list.

* Now allow h.delete_section(sec=pysec) for PythonSection.

Note that the PythonSection retains its reference to the now invalid
NEURON internal Section and former, now zombie PythonSection
is freed when the PythonSection refcnt goes to 0. Of course the internal
invalid NEURON Section is freed when it's refcount goes to 0 (in usual
circomstances, when the zombie PythonSection is destroyed.)
SectionRef.unname() and SectionRef.rename() do nothing and return 0.0
when SectionRef is referenceing a PythonSection.

* Test updated to reflect slight change in round off error of segment iterator.

* Yet another implementation of iternext_sl.

Take into account the possibility that the section returned by iternext_sl
is passed to h.delete_section.