RangeVarPlot.color keeps track of current color.
Calling the .color() method returns the previously selected color;
this method may now be called with 0 arguments (in which case it
does not change the color), or with 1 argument (in which case it
changes the color and returns the old color).

This is necessary to allow graphical tools other than InterViews
to use the chosen RangeVarPlot's color.