Multiple versions of MPI supported with -DNRN_MPI_DYNAMIC="...;..."
Only openmpi and mpich (msmpi on windows).
const char* path_prefix_to_libnrniv() returns /path/to/ the libnrniv.so
or dylib on the target machine (note the trailing '/')
or else "" if it cannot determine the path prefix.

Typo in a statement for DARWIN.

cmake mingw mpi dynamic has to link msmpi.dll to libnrnmpi_msmpi.dll