py shim: xcrun must bring all original args
NOTE: this is a bug of distutils which only takes the text
of the config_var("CC").
For the moment we patch for xcrun which is a common case in newer OSX
but it can potentially happen with other wrappers that require args.
Truth being said, xcrun is nuts as requires specifying optional arguments