On hoc error, avoid oc_restore_code tobj_count... messages.
These imply memory leaks of OBJECTTMP objects on the stack.
Also was leaking localobj objects on hoc errors.

Fixes #513

Simplest example:
>>> from neuron import h
>>> h.Vector(1).plot(1)
bad stack access: expecting (Object **); really (double)
NEURON: interpreter stack type error
 near line 0
 objref hoc_obj_[2]
                   ^
        Vector[0].plot(1)
oc_restore_code tobj_count=1 should be 0

This fixes most but not all practical sources of this error. This commit
provides a pattern to follow to provide sufficient information for the
factored out src/oc/code.c static void frame_clean_on_err(Frame* f) to
make sure that on a hoc_execerror all OBJECTTMP resources are unreffed.
In testing it is observed that in hoc calling into python that there are
many places in the python code that error handling on the python side
needs review to avoid extra python exceptions with a first line of
'The above exception was the direct cause of the following exception'.
Fixing those is deferred to a later changeset in order to be able to
highlight the major change to error handling in hoc that this represents.