Revert "On hoc error, avoid oc_restore_code tobj_count... messages. (#527)"
This reverts commit 30567cd4c4995d2abd84515c1c12bb7b60fa4e02.

It became very complicated to deal with all the edge cases that leaked
OBJECTTMP objects on the stack during error recovery. And with the general
method chosen, I was unable to fix a tobj_count message when dealing with
the SECTIONREF handling of ob.secref.parent when there was no parent or when
the secref.sec had been deleted.
Time to attempt another strategy.