On hoc error, avoid oc_restore_code tobj_count... messages.
This much simpler method replaces the reverted changeset 30567cd4c.

tobj_count messaged during an error imply memory leaks of OBJECTTMP
objects on the stack. Also was leaking localobj objects on hoc errors.

    Simplest example:
    >>> from neuron import h
    >>> h.Vector(1).plot(1)
    bad stack access: expecting (Object **); really (double)
    NEURON: interpreter stack type error
     near line 0
     objref hoc_obj_[2]
                       ^
            Vector[0].plot(1)
    oc_restore_code tobj_count=1 should be 0

I believe this fixes all sources of this error. The method on a hoc error
is to search the stack for OBJECTTMP items, unreference them and decrement
tobj_count. It is conceivable that errors during this recovery could cause
problems or if the stack is corrupt in other ways. If this happens it will
be straightforward to analyse and modify the recovery to fix the issue.