On hoc error recovery, check for duplicate unref of same stack index.
When hoc_stkobj_unref is called, the stack location is marked
STKOBJ_UNREF. The notice:
  OBJECTTMP at stack index %ld already unreffed
is printed if hoc_stkobj_unref is called again with the same stack index.
This is generally benign since this index will not be unreffed again or
tobj_count decremented. However a developer may wish to ascertain
the reason for the duplicate call to fix and to verify that there is no
unanticipated serious implication.