Tobjcount (#545)
* Revert "On hoc error, avoid oc_restore_code tobj_count... messages. (#527)"

This reverts commit 30567cd4c4995d2abd84515c1c12bb7b60fa4e02.

It became very complicated to deal with all the edge cases that leaked
OBJECTTMP objects on the stack during error recovery. And with the general
method chosen, I was unable to fix a tobj_count message when dealing with
the SECTIONREF handling of ob.secref.parent when there was no parent or when
the secref.sec had been deleted.
Time to attempt another strategy.

* On hoc error, avoid oc_restore_code tobj_count... messages.

This much simpler method replaces the reverted changeset 30567cd4c.

tobj_count messaged during an error imply memory leaks of OBJECTTMP
objects on the stack. Also was leaking localobj objects on hoc errors.

    Simplest example:
    >>> from neuron import h
    >>> h.Vector(1).plot(1)
    bad stack access: expecting (Object **); really (double)
    NEURON: interpreter stack type error
     near line 0
     objref hoc_obj_[2]
                       ^
            Vector[0].plot(1)
    oc_restore_code tobj_count=1 should be 0

I believe this fixes all sources of this error. The method on a hoc error
is to search the stack for OBJECTTMP items, unreference them and decrement
tobj_count. It is conceivable that errors during this recovery could cause
problems or if the stack is corrupt in other ways. If this happens it will
be straightforward to analyse and modify the recovery to fix the issue.

* Assignment to $o only allowed if caller arg was declared as objref.

Following used to give a segmentation violation.
from neuron import h
h('''
proc foo() {
    $o1 = new Vector()
}
''')
h.foo(None)

* hpoasgn on error does PyErr_Print and then hoc_execerror.

Prevents extra system error message from python if the hoc statement
  po.name = ... fails when po is a PythonObject.

* Bugs revealed by valgrind.

Stack tobj_count error recovery started one item too high.
nautoobj sometimes unintialized.

* valgrind noticed a memory leak when error in hoc_objgetattr.

* Hoc stack value type test must be called with valid stack pointer.

Now that hoc error recovery visits the stack to recover OBJECTTMP
resources, if tskchk calls execerror, stackp must be valid. Thus, if the
former pop idiom (tstkchk((--stackp).i, NUMBER), (--stackp).val) results
in an execerror inside tstkchk, the error recovery is not successful.

* On hoc error recovery, check for duplicate unref of same stack index.

When hoc_stkobj_unref is called, the stack location is marked
STKOBJ_UNREF. The notice:
  OBJECTTMP at stack index %ld already unreffed
is printed if hoc_stkobj_unref is called again with the same stack index.
This is generally benign since this index will not be unreffed again or
tobj_count decremented. However a developer may wish to ascertain
the reason for the duplicate call to fix and to verify that there is no
unanticipated serious implication.

* tobj_count error for bad stack access: (... really Object *)

Must do tstkchk before popping stack.