Add erase method for non-consecutive nodes in vectors of node pointers
Every node variable that is a vector of pointers to other nodes has a method to erase non-consecutive elements passed as elements of an unordered_set (generated by jinja2 and python). This effectively generalizes the procedure implemented in erase_statement_block making the old routine (erase_statement_block) obsolete. Thus, it can replace erase_statement_block and can be tested by the same unit tests that were used for testing erase_statement_block. 

It can be used to simplify the PR #230 and addresses #340