ion style error
nrn_wrote_conc was using the cinit slot rather than einit slot to
determine whether to re-initialize the reversal potential using the
nernst equation due to a change of concentation in the INITIAL block of
a mechanism that WRITE a concentation. (By default, the cinit and
einit slots have the same value but can differ if the user specified
an explicit ion_style.