`SingletonRandomString` class is now reentrant
the `instance` static method could swith the `use_numbers` member variable of the singleton
which could alter other part of the code that kept the shared pointer.
Now the `use_numbers` parameter is a parameter of `reset_random_string` member function.

This change doesn't fix an actual bug but prevent future one.