Allow NetSendBuffer_t to grow (BlueBrain/CoreNeuron#395)
Until now NetSendBuffer_t was allocated at a fixed size that could not
be changed. When using buffered net_send and invoking multiple
net_send's in INITIAL blocks in NMODL it may happen that we run out of
buffer. Currently this case is not handled and causes a buffer overflow.

As a first step NetSendBuffer_t has been given a constructor and
destructor as well as a grow function that doubles all internal buffer
sizes and copies data to the new buffers.

- By inheriting from MemoryManaged the class overloads new and free
operators to use cudaMallocManaged alloctors when necessary.
- Check for OpenACC and throw an error, as reallocation will not work
  on device

CoreNEURON Repo SHA: BlueBrain/CoreNeuron@639e9e0485f62816a5cd93938079f6297381aea0