Grow NetSendBuffer_t dynamically when necessary (#407)
The `NetSendBuffer_t` struct in coreneuron has been updated to support dynamic resizing of its internal buffers. It offers now the `grow()` function will double the buffers' size. Until now, the `net_send` buffer was allocated statically at intialization and could not be grown. When many `net_send` events are requested before they have been processed, this fixed buffer size has lead to coreneuron either crashing due to buffer overflows (when used with mod2c) or terminating with an error message (when used with nmodl). We can now call the `grow()` function to dynamically enlarge the buffers when we run out of space.