Add support for for_netcon (#410)
* Add support for for_netcon (issue #390)

For_netcon applies the operations of the block to the weights of
all the netcons. Since all the weights are on the same vector,
weights, this means appling the same operations to a mask of values
with the appropriate offset. The mask is the arguments of the
for_netcon block.

Note: the local variables are renamed to the correct position
in the corenrn vectors (no new variables added). Some of these
local variables may have been already substituted with special
characters that break rename_visitor (i.e. (*w) ). Thus, we need
to sanitize first the regex names that we are going to substitute.

- Tested with the test in /nrn/test/coreneuron/test_fornetcon.py
SoA and AoS

* Fix for AoS

Tested with /nrn/test/coreneuron/test_fornetcon.py with aos flags
(for both nmodl and coreneuron)

TODO: CI testing

* Update src/codegen/codegen_c_visitor.cpp

Co-authored-by: Pramod Kumbhar <pramod.kumbhar@epfl.ch>

* Update src/codegen/codegen_c_visitor.cpp

Co-authored-by: Pramod Kumbhar <pramod.kumbhar@epfl.ch>

* squash! simplify printing in for_netcon loop

add for_netcon_start
add for_netcon_end

* Avoid semicolon for FOR_NETCON node type
Avoid extra newline
  - for netcon is visited while printing the statement
  - instead of newline, continue writing text as
    indentation etc are already setup

Co-authored-by: Pramod Kumbhar <pramod.kumbhar@epfl.ch>
Co-authored-by: Pramod Kumbhar <pramod.s.kumbhar@gmail.com>