Add compatibility workaround for Spack's Python.
https://github.com/spack/spack/pull/16856 (merged 06/2020) include
patches to Python that cause inconsistent behaviour with the upstream
version. The patches also introduce a new `LDCXXSHARED` environment
variable that addresses a long-standing upstream issue
(https://bugs.python.org/issue1222585).

This change explicitly specifies that the PGI linker should be used when
the PGI compiler is used. This has to be specified separately for C++
and C for compatibility with the Spack patches that introduce the
LDCXXSHARED environment variable.

This avoids compilation errors due to passing PGI commandline options to
gcc, which was used as the default linker. Without the Spack patches to
Python this change was not needed, and the linker seemed to be derived
from the CC and CXX variables.