sympy replace inplace (#495)
* Add sympy_replace_solution_visitor

The visitor addresses diff_equations, linEquation, nonLinEquation. It receives 
the ast from sympy_solver_visitor. From python it receives:

- the pre_solve_statements
- the tmp_statements (only if --cse is selected)
- the solution_statements

In possibly two runs it tries to replace the solutions in place updating the
tmp_statements when needed. In case of big systems the Newton method
is employed and a system J x = f must be repeatedly solved. In this case the 
visitor must provide the J matrix and f vector. It still tries to replace the 
solutions line by line:

1 row of the J matrix and one element of the f vector.

More details in the visitor doxygen