Replace MODFILE_DIRECTORY with MODFILE_PATTERNS.
Minor change to the API for the nrn_add_test*(...) CMake helper
functions, which now take a list of relative path glob patterns matching
.mod files that should be compiled. Previously a single directory was
passed and the helper code assumed that all .mod files in that directory
should be compiled.

Also add missing documentation for the COMMAND keyword argument to
nrn_add_test.