sympy var clashes (#518)
- Remove second version of "suffix_random_string" with default param
- old_x with unique prefix to avoid clashes
- suffix_random_string: check if new_string is prefix of an existing one
- remove _make_unique_prefix and its tests as it is useless
- tmp_* is now the default to better distinguish the iterator
- Make tmp_* variables unique

We leverage the fact that suffix_random_string now returns
a prefix that is unique to create various unique strings in
ode.py appending the usual iterator. The tmp_unique_prefix
is provided to solve_lin_system. Only in case do_cse is True
we use the tmp_unique_prefix to generate the temp variables.