Fix linking against external CoreNEURON libraries.
When CoreNEURON is built as a submodule NEURON takes the variable
CORENEURON_LIB_LINK_FLAGS from a global property of the same name. If
CoreNEURON is built as an external library this variable will [soon] be
exported in coreneuron-config.cmake and should not be overwritten by the
(empty) property value.