Fix repr() and str() for AST python wrappers (#700)
The python bindings for AST classes were fixed to now properly print a
compact json representation of the node (and its children) when the repr
is called (implicitly or explictely). Additionally I added a `__str__`
to each AST node binding that will be used for string conversions (as in
print). This function returns the NMODL representation of the node (and
its children).