NEURON and CoreNEURON manage PreSyn.send differently.
With multiple threads, NEURON PreSyn.send inserts PreSyn events on
every thread event queue. For queue transfer to CoreNEURON, PreSyn
events need to be ignored if the PreSyn thread is not the queue thread.
Multiple threads + MPI still has a bug, probably due to the fact that
the NEURON equivalent to InputPreSyn is not explicitly associated with
a thread.