NEURON <-> CoreNEURON queue transfer fixes. (#1535)
* Need NetParEvents with threads and/or MPI.

On return from CoreNEURON psolve and continuing with another psolve
on NEURON, multiple threads and/or MPI require NetParEvents for
synchronization every minimum NetCon delay integration interval.

* NEURON and CoreNEURON manage PreSyn.send differently.

With multiple threads, NEURON PreSyn.send inserts PreSyn events on
every thread event queue. For queue transfer to CoreNEURON, PreSyn
events need to be ignored if the PreSyn thread is not the queue thread.
Note: The NEURON PreSyn equivalent to InputPreSyn is assoicated with
thread 0.