SaveState supports rxd (#1586)
Adds the function `register_savestate(id_, store, restore)` to the `neuron` module top-level namespace that can be used to register SaveState extensions. `store` returns a bytestring and `restore` takes a bytestring. The `id_` is used to map extension data to the right restore function. Only extensions having non-empty stored data need to be available to restore the SaveState data.

If no extension types generate state data to store, then a version 6.0 SaveState file is generated, which is backwards compatible with prior versions of NEURON. Otherwise, a version 7.0 file is generated, which is the same as 6.0 except for the version flag and the presence of state data.

The reaction-diffusion (rxd) mechanisms use this extension mechanism to allow `h.SaveState` to save and restore rxd state. For rxd, besides a version header and length, the rest of the data is stored gzip-compressed by `Species`. Additionally, new functions `rxd.save_state` and `rxd.restore_state` allow saving and restoring rxd state to/from bytestrings independently of the rest of the simulator's state, and each `rxd.Species` now has a `_state` property that can be used for reading/writing state with some error checking (e.g. size). Note: rxd only registers its savestate handlers when an `rxd.Species` is created.

Co-authored-by: Craig Kelley <craig.f.kelley@gmail.com>