partialPivLu: rework OpenMP workaround (#805)
The explicitly named partialPivLuN(...) functions still seem to be
required, but by providing specialisations of partialPivLu<N>(...) in
the header then we seem to be able to revert one change to the code
generation and allow nmodl::newton::newton_solver(...) to be used in code
compiled with OpenMP. This includes the reduced dentate model after the
changes made in #804. See also: CoreNEURON#767.