address numpy deprecation warning (#1669)
Running an rxd simulation with recent versions of numpy led to warnings like

__init__.py:735: DeprecationWarning: `np.float` is a deprecated alias for the builtin `float`. To silence this warning, use `float` by itself. Doing this will not modify any behavior and is safe. If you specifically wanted the numpy scalar type, use `np.float64` here. 
 Deprecated in NumPy 1.20; for more details and guidance: https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations 
   return numpy.ndarray((size,), numpy.float, cbuffer, order='C')
This change set standardizes the pure-Python (but not Cython) library to always use float instead of numpy.float, numpy.float_, or numpy.float64. Cython is different because our Cython code doesn't ever use any of those, but it does have a numpy.float_t that's used for defining data types but isn't actually defined in numpy.

This was used by rxd in a number of places, by the neuron.numpy_from_pointer function which makes Vector.as_numpy work, and by nonvint_block_supervisor.numpy_from_pointer. (The latter is a copy of the function from neuron and these should be merged but that's a code cleanup that goes beyond simply addressing the deprecation warning.)