disable forked for rx3d
There seem to be memory corruption issues in CVODE that are only
appearing when we run in pytest forked mode and that are thrown in the
rx3d tests. There doesn't seem to be an issue in the way we are testing
itself (maybe the fixtures in rx3d will need revisiting) but the way the
test is run now makes those underlying issues appear.
To not block adding pytest-forked into our workflow we simply disable it
in this PR for rx3d and reenable it once those CVODE issues have been
resolved and the tests pass.
Also, removed the basic test that should always crash pytest without
forked.