Make sure tests are run with actual python exe
On mac in some setups we end up still in the situation where some python
tests are executed with the python shim. This causes sanitizer libraries
to be lost in the chain of executions once again.
This change forwards the Python executable that was determined using the
SanitizerHelper.cmake utility all the way down to the subprocess
invokations inside some of the python tests.
Also, update coding conventions to fix an issue where the non-Apple
clang was not properly recognized (for sanitizer libraries).