Move virtualenv job into main build job
We'd rather not rely on jobs running on the same runner, but virtualenvs
should not be moved.
While folding jobs, we might as well set the environment variables in
the main build job as well, as it was setting several already.
This also allows us to get rid of the GIT_STRATEGY variable.