standardize string formatting, other minor doc cleanups (#2274)
Rationale: since we have version-specific NEURON documentation and we no longer support anything before Python 3.7, we can use f-strings.

There are a handful of remaining uses of % where it was used for non-trivial formatting that I left in place. .format should be gone entirely.

Also removed a few imports of gui and moved .record to the creation of the Vector. Replaced some "hh" with h.hh which allows introspection; likewise for pas. Also, replaced some uses of indgen with range.