data_handle: improve optional field support
Expand tests for optional fields in soa<...> containers, in particular
using fast_imem fields in the Node data container. Because these can be
toggled on and off, there are additional states that can be attained.

As part of this, change the pretty-printing code to use the "T**"
pointers to where the "T*" pointers are stored, instead of the "T*"
values (== vec.data()), when searching for information about the
container a data handle points into. This means that data handles to
disabled columns can pretty-print as such, instead of having to
pretty-print as being handles to unknown containers.