Fix LIKWID related build instructions.
The problems are:
* The default LIWKID target doesn't set any include directories.

The solution is to define a imported target `nrn_likwid` and manually
set the required properies. The target must be "IMPORTED" because the
Makefile for creating `special` (via NMODL) needs to extract `-I` and
`-L` commands. It does do differently for `IMPORTED` and non-IMPORTED
targets; and LIKWID needs to go the IMPORTED path.