Fix LIKWID related build instructions. (#2478)
The problem is that the default LIWKID target doesn't set
any include directories.

The solution is to define an imported target `nrn_likwid` and manually
set the required properties. The target must be "IMPORTED" because the
Makefile for creating `special` (via NMODL) needs to extract `-I` and
`-L` commands. It does do differently for `IMPORTED` and non-IMPORTED
targets, and LIKWID needs to go the IMPORTED path.