Add more REQUIRE empty checks. (#2498)
Since we need to test via a global, bugs from prior tests may affect
later tests. One variant is that if there's a bug in releasing rows, the
model wont be in the expected (empty) state and fail on an unrelated
test.

This commit adds a few more checks to make sure it's in the expected
state.