Malloc -> new[] in `grids.?pp`.
The grids related code for RxD contained the following patter:

   ptr = (data_handle<double>*) malloc(data_handle<double>*n);

The issue is, that `data_handle` needs to be initialized properly,
because the `non_owning_identifier_...` contains an `std::shared_ptr`
rather than a raw pointer.

Simply `malloc`ing a data handle causes the underlying shared pointer to
be in a random state which cause dereferencing or deallocating a random
address.