Clean checking of standard environment via CMake instead of custom checks (#2558)
In the old days C standard was not defined or supported by compilers,
so there was a lot of checks to know if such basic function was available
or such header. Today, most of them are in the standard and needed to be
here. If not the compiler is not a valid one. Get rid of all those checks to
simplify paths in the code.