Soma contour centroid oriented so 1 end is largest positive direction. (#2491)
A temporary pull request to allow comparison with #2470 

This was an attempt to see if this produces the same result as #2470.
But the conclusion was that this is not easy. We are still merging this
PR/change because we thought it would be helpful to have this change
into the NEURON master branch as a reference point if we ever have
to compare this change in isolation.