Regenerated code if code-generator changes (#2734)
While development, the code-generator might be changed. This could
result in a different `.cpp` file. However, in the make file snippet
generated by `nrnivmodl` the generated `.cpp` file doesn't have a
dependency on `nocmodl`. This means if it changes the `.cpp` doesn't need to
be recreated. Since `nrnivmodl` is called from `CMakeLists.txt` this can
cause broken builds, where changes to the source files aren't reflected
in the generated binaries.