Fail for invalid `<<` reaction equation. (BlueBrain/nmodl#1204)
The previous behaviour for

    KINETIC states {
         ~ x + y << 42.0
    }

was to print a warning and continue as if the line didn't exist. The new
behaviour is to throw an error, because it's unclear why ignoring the
line could ever lead to a correct translation of the MOD file.

NMODL Repo SHA: BlueBrain/nmodl@9292c18027154292984bc75095081a107fffdb12