Test to expose bug in array variables.
This test exposes a bug in how data is copied between NEURON and
CoreNEURON.

It must include (at least) one MOD file with array range variables. It
must check that the copy of these array range variables received by
CoreNEURON is correct.

The suspected bug is that an array range variable with M elements is
sometimes 1 and sometimes M "floating point variables"; and somehow this
is done inconsistently.