Fix CoreNEURON bug with array variables. (#2779)
* Test to expose bug in array variables.

This test exposes a bug in how data is copied between NEURON and
CoreNEURON.

It must include (at least) one MOD file with array range variables. It
must check that the copy of these array range variables received by
CoreNEURON is correct.

The suspected bug is that an array range variable with M elements is
sometimes 1 and sometimes M "floating point variables"; and somehow this
is done inconsistently.

* Store number of doubles in `nrn_prop_param_size_`.

Change `register_data_fields` to compute the number of doubles
required by a mechanism, and store that number, as opposed to the number
of variables/fields in the SoA, in the global `nrn_prop_param_size_`.

* Transfer `array_dims` to CoreNEURON.

Either via `bbcore_mech.dat` or directly via the stringstream.

* Implement CNRN->NRN SoAoS copy.
* Fix transpose during NRN->CoreNEURON (file_mode).
* Fix transpose during NRN->CoreNEURON (in-memory).

* Implement `legacy2soaos_index` and `soaos2cnrn_index`.

The two function convert legacy indices to CoreNEURON indices, i.e. the
index of the value inside the CoreNEURON datastructures.

---------

Co-authored-by: Pramod Kumbhar <pramod.s.kumbhar@gmail.com>
Co-authored-by: Michael Hines <michael.hines@yale.edu>