Improve wheel-building experience on multi-arch MacOS (#3038)
The build_static_readline_osx.bash script is a bit inflexible and cumbersome to use if trying to perform NEURON builds on both arm64 and x86_64; for instance, it installs everything to /opt/nrnwheel, which means that it doesn't really differentiate between arches, so if one wants to build a wheel, one has to remember to (re)install the readline library for the correct arch, otherwise we end up with many warnings about incompatible archs, and eventually the linker complaining about missing symbols.
This PR basically improves this by moving the install location to /opt/nrnwheel/[ARCH], where [ARCH] is detected the usual way. This means we can use build_static_readline_osx.bash on either arch, and it won't clobber any existing install of the other arch. One can also specify a custom install location by passing a single argument to build_static_readline_osx.bash.

Note that Linux builds should not be affected by any of these changes.