Improve header hygiene for `cabcode.h`. (#2948)
Moves cabcode related code into a more standard structure. Essentially,
functions are declared in headers, and defined in an associated `.cpp` file
of the same name. The `.cpp` file includes the header as the first line of
code. If a different translation unit (TU) wants to use the function it
includes the header the declares that function.

Structs and classes that are used in several translation units are defined
in headers, while structs and classes that are only used for implementation
purposes only are defined in the `.cpp` and not available to other TUs.

Explicitly writing out a function declaration (with or without `extern`) to get
access to function is discouraged.