Set the voltage from `_setdata`. (#3066)
* Move `_setdata` code in generated file.

We'll need access to `v` later. This pushes the definition of
`_setdata` further down the generated file; and prints a
declaration at the previous location. The declaration is needed
to register the function `_hoc_setdata`.

* Set the voltage from `_setdata`.

In non-VECTORIZED files, the voltage is a static double. When calling
`setdata` several other prop related global variables are set. However,
not the voltage. This commits fixes the bug and also sets `v` to the
voltage of the node associated with the prop provided.