Modernize DECREF (part 10). (#3240)
This commit refactors a part of the NRN Python bindings to use `nanobind`
objects manual reference counting. Early parts of the refactoring sequence
focused on DECREF. We've moved on to also modernize other reference
counting logic. The purpose is to address the risk of leaking when exceptions
are thrown and slowly work towards complete RAII-style reference counting.

As part of the refactoring, if needed, the scope of certain variables might be
reduced or given a new name. Additionally, NULL pointers are replaced with
`nullptr`.

This commit doesn't intentionally change reference counts.