Fix leak in `pysec_children` and `pysec_subtree`. (#3255)
The leak happens when and error occurs in the `*1` "helper" function. In
that case the `sl` (created in the top-level function) isn't DECREFed,
and also not returned. Therefore, it's leaked.

In the first case this is fixed by inlining the helper function.

In the second case, we check if the helper returns a nullptr and if not
we return the new reference to the newly created list (by using
`release().ptr()`). Note, that on the error path the `sl` will be
cleaned up due to RAII semantics.