Refactor `Py2NRNString`. (#3249)
The `Py2NRNString` consisted of (what's now) a `unique_cstr`. The ctor would
create the cstr, and `{g,s}et_pyerr` were non-static methods. They needed to
check if the cstr as a `nullptr`, but where only ever called after checking that
it was a `nullptr`.

This commit separates several notions:
  * Ownership is handled by `unique_cstr`.
  * An ASCII string is extracted from a Python object using `as_ascii`.
  * Both `{g,s}et_pyerr` were made static an are independent of each other
    and from converting a Python object to an (ASCII) C-string.
  * The `PyErr2NRNString` was removed.