CoreNrn: support for rebalanced files.dat (#3147)
Context: For MMB (BBPP134-917) we need to be able to load a set
of .dat files distributed in given way. However for the moment
CoreNeuron only reads it in plain RoundRobin, which is not suitable
to load ranks a different number of files.

Changes: We introduce a mark value (-1 in this case) indicating no-data
so that, even with round-robin, we can define precisely which rank will
load a given data file. All ranks will be able to jump strides of size
`N_RANKS` (at their respective offset) and find entries targeted at them.
Upon seeing a `-1` the rank knows it is done and can stop reading.

This is a minor addition which enabled us to load balance and run the
MMB simulation.