Partially remove CLI args to `setup.py`
Instead of passing `--enable-X` or `--disable-Y` or `--without-Z` to
`setup.py`, it makes more sense to just directly use the CMake
variables. The following are now set as env variables in
`build_wheels.bash`:

- NRN_ENABLE_PYTHON_DYNAMIC
- NRN_RX3D_OPT_LEVEL
- NRN_ENABLE_RX3D
- NRN_ENABLE_INTERVIEWS
- NRN_ENABLE_MPI
- NRN_ENABLE_MUSIC
- NRN_ENABLE_CORENEURON

This eliminates the need for the following args:

- without-nrnpython
- rx3d-opt-level
- disable-rx3d
- disable-iv
- disable-mpi
- enable-music
- enable-coreneuron

This is useful since we want to replace calls of `python setup.py
[args]` with `pip install`, i.e. towards conformance with PEP 517 (see
also: #2696).

It also refactors `build_wheels.bash` to make it more modular and
using the "don't repeat yourself" (DRY) principle.

The user-facing changes are as follows:

- `build_wheels.bash` can now be ran without _any_ parameters because
  the platform is automatically detected (using `uname -s`)
- the possible values of the first argument to `build_wheels.bash` were
  changed from `linux` to `Linux`, and `osx` to `Darwin` to conform to
  the output to `uname -s`. The special value `CI` remains valid.